feat(store)!: one .build() terminal — unify EventStore/ZeroCopyEventStore (#248)
The owning-vs-borrowing codec distinction leaked into the builder as two
terminals (build / build_zero_copy) and two facade types. The two facades
were not behaviorally distinct: identical (encode-only) save paths, and load
paths differing by one line (replay(&event) vs replay(event)), both feeding
AggregateRoot::replay(&EventOf<A>).

Collapse to a single EventStore<S,C,A> whose load path is bounded on
`for<'a> Decode::Output<'a>: Borrow<EventOf<A>>` and calls out.borrow() — the
std Borrow blanket impls (Borrow<T> for T and for &T) unify Output = E and
Output = &E. ZeroCopyEventStore, build_zero_copy, and save_borrowing deleted.
The choice is now inferred from the codec, not restated at the call site
(serde DeserializeOwned precedent).

No capability change: the old ZeroCopyEventStore bound was Output = &EventOf<A>
literally, so Borrow<EventOf<A>> covers exactly the same {E, &E} set.
